Changing to consume using modules. Fixing/adding package.json files
