+++ org.OpenT2T.Onboarding.WinkHub/v2/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "opent2t-onboarding-winkhub",
+  "name": "opent2t-onboarding-winkhub-v2",
   "version": "1.0.0",
   "description": "Wink Hub onboarding module for Open Translators to Things",
   "author": "OpenT2T",
@@ -22,6 +22,13 @@
     "jshint": "^2.9.2",
     "assert": "^1.4.1"
   },
+  "main": "winkhubtranslator.js",
+  "files": [
+    "winkonboarding.js",
+    "winkhubtranslator.js",
+    "winkhub.js",
+    "package.json"
+  ],
   "opent2t": {
     "interfaces": {
       "winkhub": {
