+++ org.OpenT2T.Onboarding.WinkHub/v2/winkhub.js
@@ -2,5 +2,5 @@
 var AllJoynConverter = require("openT2T/converters/AllJoynConverter");
 
 module.exports = AllJoynConverter.readDeviceInterfacesFromFile(
-    require("path").join(__dirname, "../../org.OpenT2T.Onboarding.Hub/org.OpenT2T.Sample.SuperPopular.Hub.xml"))[0];
+    require("path").join(__dirname, "org.OpenT2T.Sample.SuperPopular.Hub.xml"))[0];
